requirements.txt
streamlit==1.52.0
langchain==0.3.27
langchain-openai==0.3.35
langgraph==0.6.6
openai==2.2.0
python-dotenv==1.2.1